Move psconvert to a separate file
