+++ pygmt/src/__init__.py
@@ -40,6 +40,7 @@ from pygmt.src.nearneighbor import nearneighbor
 from pygmt.src.plot import plot
 from pygmt.src.plot3d import plot3d
 from pygmt.src.project import project
+from pygmt.src.psconvert import psconvert
 from pygmt.src.rose import rose
 from pygmt.src.select import select
 from pygmt.src.shift_origin import shift_origin
